+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 -   SC-46: `bm server serverless --deployment-bucket` flag to specify a custom AWS S3 bucket for deployment
 
+-   SC-46: `bm server serverless --execution-role` flag to specify a custom AWS IAM role ARN during function execution
+
 
 ### Fixed